utils: Sharing Expiration with main codebase
